Update to latest openbmc/pldm commit 3d34b2c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/125406
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/subtree/openbmc/pldm/libpldmresponder/base.cpp
@@ -217,7 +217,7 @@ void Handler::processSetEventReceiver(
         if (rc || completionCode)
         {
             std::cerr << "Failed to decode setEventReceiver command response,"
-                      << " rc=" << rc << "cc=" << (uint8_t)completionCode
+                      << " rc=" << rc << "cc=" << (unsigned)completionCode
                       << "\n";
             pldm::utils::reportError(
                 "xyz.openbmc_project.bmc.pldm.InternalFailure");
